+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Made SVG icon titles more consistent
+- Migrated style guide build to use Pattern Lab 3
 - Updated various dependencies
 - Travis build now uses Node version specified in .nvmrc